Fix Google auth.
I apparently didn't realize that the GoogleSignIn stuff was just doing Google
OAuth auth, not actually doing the bridging to Firebase that turns out to be a
separate step. So now when we auth with Google, we then turn around and send
those credentials to Firebase, which gives us a proper Firebase user id (which
now matches the one we get on the web, yay).

The passages, they twist.