Switch to Material widgets from Cupertino.
I was hoping this would fix the weird everything takes two taps bug, but alas
not entirely. There's something broken deep in the bowels of Flutter that I'll
just have to wait for them to fix.

There's also some weird bug in the bowels of MobX that is preventing profile
images/names from updating when they are resolved asynchronously. Sigh. I'm
going to stop poking it for now and just ship this since it's a bit better than
what we have and it handles npc profiles.